feat(Coordinate): move to its own module
