+++ libs/navigation.py
@@ -6,29 +6,8 @@ from loguru import logger
 
 from libs import rover
 from libs.config import Config
-from libs.gps_controller import GpsController
-from gps_controller import calculate_true_bearing
-
-
-@dataclass()
-class Coordinate:
-    latitude: float
-    longitude: float
-
-    def get_latitude(self) -> float:
-        self.latitude
-
-    def get_longitude(self) -> float:
-        self.longitude
-
-    def set_latitude(self, latitude: float):
-        self.latitude = latitude
-
-    def set_longitude(self, longitude: float):
-        self.longitude = longitude
-
-    def __str__(self):
-        return f"({self.latitude}, {self.longitude})"
+from libs.coordinate import Coordinate
+from libs.gps_controller import GpsController, calculate_true_bearing, GpsReport
 
 
 @dataclass()
